feat: create new order

+++ Front-End/src/pages/ProductPages/Cart.jsx
@@ -1,22 +1,55 @@
-import React from "react";
+import React, { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { Link } from "react-router-dom";
+import Swal from "sweetalert2";
 
 import {
   removeAllFromCart,
   removeFromCart,
   updateQuantity,
 } from "./../../store/Slice/cart";
+import { userPostNewOrder } from "../../api/apiOrder";
 export default function Cart() {
   const cartProducts = useSelector((state) => state.cart.cartList);
   const dispatch = useDispatch();
 
-  function addNewOrder() {
-    dispatch(removeAllFromCart());
-    // Axios.post()
+  let totalPrice =
+    cartProducts.reduce(
+      (accumulator, currentValue) =>
+        accumulator + currentValue.quantity * currentValue.price,
+      0
+    ) + 10;
+
+  async function addNewOrder() {
+    let newProducts = cartProducts.map((product) => ({
+      _id: product._id,
+      name: product.name,
+      price: product.price,
+      quantity: product.quantity,
+    }));
+    let data = await userPostNewOrder({
+      products: newProducts,
+      total_price: totalPrice,
+      address: {
+        city: "alex",
+        street: "miami",
+        building: "55",
+      },
+      contact_phone: "01110464712",
+    });
+    console.log(data);
+    if (data) {
+      Swal.fire({
+        icon: "success",
+        title: "Order Confirmed",
+        showConfirmButton: false,
+        timer: 1500,
+      });
+      dispatch(removeAllFromCart());
+    }
   }
   return (
-    <section id="cart" className="section-p1 container bg-light ">
+    <section id="cart" className="section-p1 container text-light ">
       <table className="full-width">
         <thead>
           <tr>
@@ -40,30 +73,50 @@ export default function Cart() {
               <td>{product.name}</td>
               <td>${product.price}</td>
               <td>
-                <input
-                  type="number"
-                  max={product.available}
-                  placeholder={product.quantity}
-                  onChange={(e) => {
-                    // Get the entered value and the maximum value
-                    const enteredValue = +e.target.value;
-                    const maxValue = +product.available;
-
-                    // Set the value of the input field to the entered value or the maximum value
-                    let newValue = Math.min(enteredValue, maxValue);
-                    if (newValue === 0) newValue = 1;
-                    dispatch(
-                      updateQuantity({ id: product._id, quantity: newValue })
-                    );
-                  }}
-                  onKeyDown={(e) => {
-                    const maxValue = +product.available;
-                    if (+e.target.value > maxValue) {
-                      e.preventDefault();
-                    }
-                  }}
-                  min="1"
-                />
+                <div className="qty text-center">
+                  <span
+                    className="minus bg-danger"
+                    onClick={() => {
+                      let newqunatity = product.quantity;
+                      newqunatity--;
+                      if (newqunatity < 1) return;
+                      dispatch(
+                        updateQuantity({
+                          id: product._id,
+                          quantity: newqunatity,
+                        })
+                      );
+                    }}
+                  >
+                    -
+                  </span>
+                  <input
+                    type="number"
+                    className="count border-0 bg-transparent text-white "
+                    name="qty"
+                    value={product.quantity}
+                    disabled
+                    max={product.available}
+                    placeholder={product.quantity}
+                    min="1"
+                  />
+                  <span
+                    className="plus bg-success"
+                    onClick={() => {
+                      let newqunatity = product.quantity;
+                      newqunatity++;
+                      if (newqunatity > product.available) return;
+                      dispatch(
+                        updateQuantity({
+                          id: product._id,
+                          quantity: newqunatity,
+                        })
+                      );
+                    }}
+                  >
+                    +
+                  </span>
+                </div>
               </td>
               <td>{product.available - product.quantity}</td>
               <td onClick={() => dispatch(removeFromCart(product._id))}>
@@ -80,6 +133,9 @@ export default function Cart() {
       <div className="d-flex justify-content-end">
         <div className="total-price ml-auto">
           <table className="full-width">
+            <tbody>
+
+
             <tr>
               <td>SUBTOTAL</td>
               <td>
@@ -97,15 +153,9 @@ export default function Cart() {
             </tr>
             <tr>
               <td>TOTAL</td>
-              <td>
-                $
-                {cartProducts.reduce(
-                  (accumulator, currentValue) =>
-                    accumulator + currentValue.quantity * currentValue.price,
-                  0
-                ) + 10}{" "}
-              </td>
+              <td>{totalPrice}</td>
             </tr>
+            </tbody>
           </table>
           <button
             className="btn btn-outline-success w-100"
